Move stuff to source file
