+++ include/core/mortar_coupling_manager.h
@@ -1,4 +1,4 @@
-// SPDX-FileCopyrightText: Copyright (c) 2023-2024 The Lethe Authors
+// SPDX-FileCopyrightText: Copyright (c) 2025 The Lethe Authors
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception OR LGPL-2.1-or-later
 
 #ifndef lethe_core_mortar_coupling_manager_h
@@ -545,86 +545,6 @@ symm_scalar_product_add(Tensor<2, dim, Number>       &v_gradient,
       }
 }
 
-
-/**
- * @brief Compute the number of subdivisions at the rotor-stator interface and the rotor radius
- * @param[in] dof_handler DoFHandler associated to the triangulation
- * @param[in] mortar_parameters The information about the mortar method
- * control, including the rotor mesh parameters
- *
- * @return n_subdivisions Number of cells at the interface between inner
- * and outer domains
- * @return radius Radius at the interface between inner and outer domains
- */
-template <int dim>
-std::pair<unsigned int, double>
-compute_n_subdivisions_and_radius(
-  const DoFHandler<dim>         &dof_handler,
-  const Parameters::Mortar<dim> &mortar_parameters)
-{
-  // Number of subdivisions per process
-  unsigned int n_subdivisions_local = 0;
-  // Number of vertices at the boundary per process
-  unsigned int n_vertices_local = 0;
-  // Tolerance for rotor radius computation
-  const double tolerance = 1e-8;
-  // Min and max values for rotor radius computation
-  double radius_min = 1e12;
-  double radius_max = 1e-12;
-
-  // Check number of faces and vertices at the rotor-stator interface
-  for (const auto &cell :
-       dof_handler.get_triangulation().active_cell_iterators())
-    {
-      if (cell->is_locally_owned())
-        {
-          for (const auto &face : cell->face_iterators())
-            {
-              if (face->at_boundary())
-                {
-                  if (face->boundary_id() ==
-                      mortar_parameters.rotor_boundary_id)
-                    {
-                      n_subdivisions_local++;
-                      for (unsigned int vertex_index = 0;
-                           vertex_index < face->n_vertices();
-                           vertex_index++)
-                        {
-                          n_vertices_local++;
-                          auto   v = face->vertex(vertex_index);
-                          double radius_current =
-                            v.distance(mortar_parameters.center_of_rotation);
-                          radius_min = std::min(radius_min, radius_current);
-                          radius_max = std::max(radius_max, radius_current);
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-  // Total number of faces
-  const unsigned int n_subdivisions =
-    Utilities::MPI::sum(n_subdivisions_local,
-                        dof_handler.get_mpi_communicator());
-
-  // Min and max values over all processes
-  radius_min =
-    Utilities::MPI::min(radius_min, dof_handler.get_mpi_communicator());
-  radius_max =
-    Utilities::MPI::max(radius_max, dof_handler.get_mpi_communicator());
-
-  AssertThrow(
-    std::abs(radius_max - radius_min) < tolerance,
-    ExcMessage(
-      "The computed radius of the rotor mesh has a variation greater than the tolerance across the rotor domain, meaning that the prescribed center of rotation and the rotor geometry are not in accordance."));
-
-  // Final radius value
-  const double radius = radius_min;
-
-  return {n_subdivisions, radius};
-}
-
 template <int dim, typename Number>
 class CouplingOperatorBase
 {
@@ -820,30 +740,7 @@ public:
                    const unsigned int               bid_1,
                    const double                     sip_factor = 1.0,
                    const unsigned int first_selected_component = 0,
-                   const double       penalty_factor_grad      = 1.0)
-    : CouplingOperatorBase<dim, Number>(
-        mapping,
-        dof_handler,
-        constraints,
-        quadrature,
-        n_subdivisions,
-        1,
-        2 * n_components,
-        radius,
-        rotate_pi,
-        bid_0,
-        bid_1,
-        sip_factor,
-        get_relevant_dof_indices(dof_handler.get_fe(),
-                                 first_selected_component),
-        penalty_factor_grad)
-    , fe_sub(dof_handler.get_fe().base_element(
-               dof_handler.get_fe()
-                 .component_to_base_index(first_selected_component)
-                 .first),
-             n_components)
-    , phi_m(mapping, fe_sub, update_values | update_gradients)
-  {}
+                   const double       penalty_factor_grad      = 1.0);
 
   /**
    * @brief Constructor of the class
@@ -860,742 +757,32 @@ public:
                    const Quadrature<dim>           &quadrature,
                    const Parameters::Mortar<dim>   &mortar_parameters,
                    const unsigned int first_selected_component = 0,
-                   const double       penalty_factor_grad      = 1.0)
-    : CouplingOperatorBase<dim, Number>(
-        mapping,
-        dof_handler,
-        constraints,
-        quadrature,
-        compute_n_subdivisions_and_radius(dof_handler, mortar_parameters).first,
-        1,
-        2 * n_components,
-        compute_n_subdivisions_and_radius(dof_handler, mortar_parameters)
-          .second,
-        mortar_parameters.rotor_mesh->rotation_angle,
-        mortar_parameters.rotor_boundary_id,
-        mortar_parameters.stator_boundary_id,
-        mortar_parameters.sip_factor,
-        get_relevant_dof_indices(dof_handler.get_fe(),
-                                 first_selected_component),
-        penalty_factor_grad)
-    , fe_sub(dof_handler.get_fe().base_element(
-               dof_handler.get_fe()
-                 .component_to_base_index(first_selected_component)
-                 .first),
-             n_components)
-    , phi_m(mapping, fe_sub, update_values | update_gradients)
-  {}
+                   const double       penalty_factor_grad      = 1.0);
 
   static std::vector<unsigned int>
   get_relevant_dof_indices(const FiniteElement<dim> &fe,
-                           const unsigned int        first_selected_component)
-  {
-    std::vector<unsigned int> result;
-
-    for (unsigned int i = 0; i < fe.n_dofs_per_cell(); ++i)
-      if ((first_selected_component <= fe.system_to_component_index(i).first) &&
-          (fe.system_to_component_index(i).first <
-           first_selected_component + n_components))
-        result.push_back(i);
-
-    return result;
-  }
+                           const unsigned int        first_selected_component);
 
   void
-  local_reinit(const typename Triangulation<dim>::cell_iterator &cell,
-               const ArrayView<const Point<dim, Number>> &points) const override
-  {
-    this->phi_m.reinit(cell, points);
-  }
+  local_reinit(
+    const typename Triangulation<dim>::cell_iterator &cell,
+    const ArrayView<const Point<dim, Number>>        &points) const override;
 
   void
   local_evaluate(const Vector<Number> &buffer,
                  const unsigned int    ptr_q,
                  const unsigned int    q_stride,
-                 Number               *all_value_m) const override
-  {
-    this->phi_m.evaluate(buffer,
-                         EvaluationFlags::values | EvaluationFlags::gradients);
-
-    for (const auto q : this->phi_m.quadrature_point_indices())
-      {
-        const unsigned int q_index = ptr_q + q;
-
-        const auto normal     = this->all_normals[q_index];
-        const auto value_m    = this->phi_m.get_value(q);
-        const auto gradient_m = contract(this->phi_m.get_gradient(q), normal);
-
-        BufferRW<Number> buffer_m(all_value_m, q * 2 * n_components * q_stride);
-
-        buffer_m.write(value_m);
-        buffer_m.write(gradient_m);
-      }
-  }
+                 Number               *all_value_m) const override;
 
   void
   local_integrate(Vector<Number>    &buffer,
                   const unsigned int ptr_q,
                   const unsigned int q_stride,
                   Number            *all_value_m,
-                  Number            *all_value_p) const override
-  {
-    for (const auto q : this->phi_m.quadrature_point_indices())
-      {
-        const unsigned int q_index = ptr_q + q;
-
-        BufferRW<Number> buffer_m(all_value_m, q * 2 * n_components * q_stride);
-        BufferRW<Number> buffer_p(all_value_p, q * 2 * n_components * q_stride);
-
-        const auto value_m           = buffer_m.template read<value_type>();
-        const auto value_p           = buffer_p.template read<value_type>();
-        const auto normal_gradient_m = buffer_m.template read<value_type>();
-        const auto normal_gradient_p = buffer_p.template read<value_type>();
-
-        const auto JxW               = this->all_weights[q_index];
-        const auto penalty_parameter = this->all_penalty_parameter[q_index];
-        const auto normal            = this->all_normals[q_index];
-
-        const auto value_jump = (value_m - value_p);
-        const auto gradient_normal_avg =
-          (normal_gradient_m - normal_gradient_p) * 0.5;
-
-        const double sigma = penalty_parameter * this->penalty_factor;
-
-        // - (n avg(∇v), jump(u))
-        this->phi_m.submit_gradient(outer(-value_jump, normal) * 0.5 * JxW, q);
-
-        // + (jump(v), σ jump(u) - avg(∇u) n)
-        this->phi_m.submit_value((value_jump * sigma - gradient_normal_avg) *
-                                   JxW,
-                                 q);
-      }
-
-    this->phi_m.test_and_sum(buffer,
-                             EvaluationFlags::values |
-                               EvaluationFlags::gradients);
-  }
+                  Number            *all_value_p) const override;
 
   const FESystem<dim>       fe_sub;
   mutable FEPointIntegrator phi_m;
 };
-
-
-
-template <int dim, typename Number>
-CouplingOperatorBase<dim, Number>::CouplingOperatorBase(
-  const Mapping<dim>              &mapping,
-  const DoFHandler<dim>           &dof_handler,
-  const AffineConstraints<Number> &constraints,
-  const Quadrature<dim>            quadrature,
-  const unsigned int               n_subdivisions,
-  const unsigned int               n_components,
-  const unsigned int               N,
-  const double                     radius,
-  const double                     rotate_pi,
-  const unsigned int               bid_0,
-  const unsigned int               bid_1,
-  const double                     sip_factor,
-  const std::vector<unsigned int>  relevant_dof_indices,
-  const double                     penalty_factor_grad)
-  : mapping(mapping)
-  , dof_handler(dof_handler)
-  , constraints(constraints)
-  , quadrature(quadrature)
-  , n_components(n_components)
-  , N(N)
-  , bid_0(bid_0)
-  , bid_1(bid_1)
-  , relevant_dof_indices(relevant_dof_indices)
-  , n_dofs_per_cell(relevant_dof_indices.size())
-{
-  penalty_factor =
-    compute_penalty_factor(dof_handler.get_fe().degree, sip_factor);
-  this->penalty_factor_grad = penalty_factor_grad;
-
-  mortar_manager_q = std::make_shared<MortarManager<dim>>(
-    n_subdivisions, quadrature.get_tensor_basis()[0].size(), radius, rotate_pi);
-
-  mortar_manager_cell =
-    std::make_shared<MortarManager<dim>>(n_subdivisions, 1, radius, rotate_pi);
-
-  const unsigned int n_points    = mortar_manager_q->get_n_points();
-  const unsigned int n_sub_cells = mortar_manager_cell->get_n_points();
-
-  std::vector<types::global_dof_index> is_local;
-  std::vector<types::global_dof_index> is_ghost;
-  std::vector<types::global_dof_index> is_local_cell;
-  std::vector<types::global_dof_index> is_ghost_cell;
-
-  for (const auto &cell : dof_handler.active_cell_iterators())
-    if (cell->is_locally_owned())
-      for (const auto face_no : cell->face_indices())
-        if ((cell->face(face_no)->boundary_id() == bid_0) ||
-            (cell->face(face_no)->boundary_id() == bid_1))
-          {
-            const auto face = cell->face(face_no);
-
-            // indices
-            const auto indices_q =
-              mortar_manager_q->get_indices(get_rad(cell, face));
-            for (unsigned int ii = 0; ii < indices_q.size(); ++ii)
-              {
-                unsigned int i = indices_q[ii];
-                unsigned int id_local, id_ghost;
-
-                if (face->boundary_id() == bid_0)
-                  {
-                    id_local = i;
-                    id_ghost = i + n_points;
-                  }
-                else if (face->boundary_id() == bid_1)
-                  {
-                    id_local = i + n_points;
-                    id_ghost = i;
-                  }
-
-                is_local.emplace_back(id_local);
-                is_ghost.emplace_back(id_ghost);
-              }
-
-            // indices of cells/DoFs on them
-            const auto indices =
-              mortar_manager_cell->get_indices(get_rad(cell, face));
-
-            const auto local_dofs = this->get_dof_indices(cell);
-
-            for (unsigned int ii = 0; ii < indices.size(); ++ii)
-              {
-                unsigned int i = indices[ii];
-                unsigned int id_local, id_ghost;
-
-                if (face->boundary_id() == bid_0)
-                  {
-                    id_local = i;
-                    id_ghost = i + n_sub_cells;
-                  }
-                else if (face->boundary_id() == bid_1)
-                  {
-                    id_local = i + n_sub_cells;
-                    id_ghost = i;
-                  }
-
-                is_local_cell.emplace_back(id_local);
-                is_ghost_cell.emplace_back(id_ghost);
-
-                for (const auto i : local_dofs)
-                  dof_indices.emplace_back(i);
-              }
-
-            // weights
-            const auto weights =
-              mortar_manager_q->get_weights(get_rad(cell, face));
-            all_weights.insert(all_weights.end(),
-                               weights.begin(),
-                               weights.end());
-
-            // normals
-            // points (also convert real to unit coordinates)
-            if (false)
-              {
-                const auto points =
-                  mortar_manager_q->get_points(get_rad(cell, face));
-                std::vector<Point<dim, Number>> points_ref(points.size());
-                mapping.transform_points_real_to_unit_cell(cell,
-                                                           points,
-                                                           points_ref);
-                all_points_ref.insert(all_points_ref.end(),
-                                      points_ref.begin(),
-                                      points_ref.end());
-
-                std::vector<Point<dim - 1>> quad;
-
-                for (const auto p : points_ref)
-                  {
-                    if ((face_no / 2) == 0)
-                      {
-                        quad.emplace_back(p[1]);
-                      }
-                    else if ((face_no / 2) == 1)
-                      {
-                        quad.emplace_back(p[0]);
-                      }
-                    else
-                      {
-                        AssertThrow(false, ExcInternalError());
-                      }
-                  }
-
-                FEFaceValues<dim> fe_face_values(mapping,
-                                                 cell->get_fe(),
-                                                 quad,
-                                                 update_normal_vectors);
-
-                fe_face_values.reinit(cell, face_no);
-
-                all_normals.insert(all_normals.end(),
-                                   fe_face_values.get_normal_vectors().begin(),
-                                   fe_face_values.get_normal_vectors().end());
-              }
-            else
-              {
-                auto normals =
-                  mortar_manager_q->get_normals(get_rad(cell, face));
-                if (face->boundary_id() == bid_1)
-                  for (auto &normal : normals)
-                    normal *= -1.0;
-                all_normals.insert(all_normals.end(),
-                                   normals.begin(),
-                                   normals.end());
-
-                auto points =
-                  mortar_manager_q->get_points_ref(get_rad(cell, face));
-
-                const bool flip =
-                  (face->vertex(0)[0] * face->vertex(1)[1] -
-                   face->vertex(0)[1] * face->vertex(1)[0]) < 0.0;
-
-                if (flip)
-                  for (auto &p : points)
-                    p[0] = 1.0 - p[0];
-
-                if (face_no / 2 == 0)
-                  {
-                    for (auto &p : points)
-                      all_points_ref.emplace_back(face_no % 2, p[0]);
-                  }
-                else if (face_no / 2 == 1)
-                  {
-                    for (auto &p : points)
-                      all_points_ref.emplace_back(p[0], face_no % 2);
-                  }
-                else
-                  {
-                    AssertThrow(false, ExcNotImplemented());
-                  }
-              }
-
-            // penalty parmeter
-            const Number penalty_parameter = compute_penalty_parameter(cell);
-
-            for (unsigned int i = 0;
-                 i < mortar_manager_q->get_n_points(get_rad(cell, face));
-                 ++i)
-              all_penalty_parameter.emplace_back(penalty_parameter);
-          }
-
-  // setup communication
-  partitioner.reinit(is_local, is_ghost, dof_handler.get_mpi_communicator());
-  partitioner_cell.reinit(is_local_cell,
-                          is_ghost_cell,
-                          dof_handler.get_mpi_communicator());
-
-  // finalized penalty parameters
-  std::vector<Number> all_penalty_parameter_ghost(all_penalty_parameter.size());
-  partitioner.template export_to_ghosted_array<Number, 1>(
-    all_penalty_parameter, all_penalty_parameter_ghost);
-  for (unsigned int i = 0; i < all_penalty_parameter.size(); ++i)
-    all_penalty_parameter[i] =
-      std::min(all_penalty_parameter[i], all_penalty_parameter_ghost[i]);
-
-  // finialize DoF indices
-  dof_indices_ghost.resize(dof_indices.size());
-  partitioner_cell.template export_to_ghosted_array<types::global_dof_index, 0>(
-    dof_indices, dof_indices_ghost, n_dofs_per_cell);
-
-  {
-    auto locally_owned_dofs = constraints.get_locally_owned_indices();
-    auto constraints_to_make_consistent = constraints.get_local_lines();
-
-
-    for (unsigned int i = 0; i < dof_indices.size(); ++i)
-      {
-        constraints_to_make_consistent.add_index(dof_indices[i]);
-        constraints_to_make_consistent.add_index(dof_indices_ghost[i]);
-      }
-
-    constraints_extended.reinit(locally_owned_dofs,
-                                constraints_to_make_consistent);
-    constraints_extended.merge(
-      constraints,
-      AffineConstraints<Number>::MergeConflictBehavior::no_conflicts_allowed,
-      true);
-
-    constraints_extended.make_consistent_in_parallel(
-      locally_owned_dofs,
-      constraints_to_make_consistent,
-      dof_handler.get_mpi_communicator());
-  }
-}
-
-template <int dim, typename Number>
-const AffineConstraints<Number> &
-CouplingOperatorBase<dim, Number>::get_affine_constraints() const
-{
-  return constraints_extended;
-}
-
-template <int dim, typename Number>
-Number
-CouplingOperatorBase<dim, Number>::compute_penalty_factor(
-  const unsigned int degree,
-  const Number       factor) const
-{
-  return factor * (degree + 1.0) * (degree + 1.0);
-}
-
-template <int dim, typename Number>
-Number
-CouplingOperatorBase<dim, Number>::compute_penalty_parameter(
-  const typename Triangulation<dim>::cell_iterator &cell) const
-{
-  const unsigned int degree = dof_handler.get_fe().degree;
-
-  FE_Nothing<dim> fe_nothing;
-
-  dealii::QGauss<dim>   quadrature(degree + 1);
-  dealii::FEValues<dim> fe_values(mapping,
-                                  fe_nothing,
-                                  quadrature,
-                                  dealii::update_JxW_values);
-
-  dealii::QGauss<dim - 1>   face_quadrature(degree + 1);
-  dealii::FEFaceValues<dim> fe_face_values(mapping,
-                                           fe_nothing,
-                                           face_quadrature,
-                                           dealii::update_JxW_values);
-
-  fe_values.reinit(cell);
-
-  Number volume = 0;
-  for (unsigned int q = 0; q < quadrature.size(); ++q)
-    volume += fe_values.JxW(q);
-
-  Number surface_area = 0;
-  for (const auto f : cell->face_indices())
-    {
-      fe_face_values.reinit(cell, f);
-      const Number factor = 0.5; /*Assuming that we are within the domain*/
-      for (unsigned int q = 0; q < face_quadrature.size(); ++q)
-        surface_area += fe_face_values.JxW(q) * factor;
-    }
-
-  return surface_area / volume;
-}
-
-template <int dim, typename Number>
-double
-CouplingOperatorBase<dim, Number>::get_rad(
-  const typename Triangulation<dim>::cell_iterator &cell,
-  const typename Triangulation<dim>::face_iterator &face) const
-{
-  if (false)
-    return point_to_angle(face->center());
-  else
-    return point_to_angle(mapping.transform_unit_to_real_cell(
-      cell,
-      MappingQ1<dim>().transform_real_to_unit_cell(cell, face->center())));
-}
-
-template <int dim, typename Number>
-std::vector<types::global_dof_index>
-CouplingOperatorBase<dim, Number>::get_dof_indices(
-  const typename DoFHandler<dim>::active_cell_iterator &cell) const
-{
-  std::vector<types::global_dof_index> local_dofs_all(
-    dof_handler.get_fe().n_dofs_per_cell());
-  cell->get_dof_indices(local_dofs_all);
-
-  std::vector<types::global_dof_index> local_dofs(n_dofs_per_cell);
-
-  for (unsigned int i = 0; i < n_dofs_per_cell; ++i)
-    local_dofs[i] = local_dofs_all[relevant_dof_indices[i]];
-
-  return local_dofs;
-}
-
-template <int dim, typename Number>
-void
-CouplingOperatorBase<dim, Number>::vmult_add(VectorType       &dst,
-                                             const VectorType &src) const
-{
-  // 1) evaluate
-  unsigned int ptr_q = 0;
-
-  Vector<Number> buffer;
-
-  std::vector<Number> all_value_m(all_normals.size() * N);
-  std::vector<Number> all_value_p(all_normals.size() * N);
-
-  for (const auto &cell : dof_handler.active_cell_iterators())
-    if (cell->is_locally_owned())
-      for (const auto &face : cell->face_iterators())
-        if ((face->boundary_id() == bid_0) || (face->boundary_id() == bid_1))
-          {
-            const unsigned int n_q_points =
-              mortar_manager_q->get_n_points(get_rad(cell, face));
-
-            local_reinit(cell,
-                         ArrayView<const Point<dim, Number>>(
-                           all_points_ref.data() + ptr_q, n_q_points));
-
-            buffer.reinit(n_dofs_per_cell);
-
-            const auto local_dofs = this->get_dof_indices(cell);
-
-            for (unsigned int i = 0; i < local_dofs.size(); ++i)
-              buffer[i] = src[local_dofs[i]];
-
-            local_evaluate(buffer, ptr_q, 1, all_value_m.data() + ptr_q * N);
-
-            ptr_q += n_q_points;
-          }
-
-  // 2) communicate
-  partitioner.template export_to_ghosted_array<Number, 0>(
-    ArrayView<const Number>(reinterpret_cast<Number *>(all_value_m.data()),
-                            all_value_m.size() * n_components),
-    ArrayView<Number>(reinterpret_cast<Number *>(all_value_p.data()),
-                      all_value_p.size() * n_components),
-    N * n_components);
-
-  // 3) integrate
-  ptr_q = 0;
-  for (const auto &cell : dof_handler.active_cell_iterators())
-    if (cell->is_locally_owned())
-      for (const auto &face : cell->face_iterators())
-        if ((face->boundary_id() == bid_0) || (face->boundary_id() == bid_1))
-          {
-            const unsigned int n_q_points =
-              mortar_manager_q->get_n_points(get_rad(cell, face));
-
-            local_reinit(cell,
-                         ArrayView<const Point<dim, Number>>(
-                           all_points_ref.data() + ptr_q, n_q_points));
-
-            buffer.reinit(n_dofs_per_cell);
-            local_integrate(buffer,
-                            ptr_q,
-                            1,
-                            all_value_m.data() + ptr_q * N,
-                            all_value_p.data() + ptr_q * N);
-
-            const auto local_dofs = this->get_dof_indices(cell);
-            constraints.distribute_local_to_global(buffer, local_dofs, dst);
-
-            ptr_q += n_q_points;
-          }
-
-  dst.compress(VectorOperation::add);
-}
-
-template <int dim, typename Number>
-void
-CouplingOperatorBase<dim, Number>::add_diagonal_entries(
-  VectorType &diagonal) const
-{
-  unsigned int ptr_q = 0;
-
-  Vector<Number>      buffer, diagonal_local;
-  std::vector<Number> all_value_m, all_value_p;
-
-  for (const auto &cell : dof_handler.active_cell_iterators())
-    if (cell->is_locally_owned())
-      for (const auto &face : cell->face_iterators())
-        if ((face->boundary_id() == bid_0) || (face->boundary_id() == bid_1))
-          {
-            const unsigned int n_q_points =
-              mortar_manager_q->get_n_points(get_rad(cell, face));
-
-            local_reinit(cell,
-                         ArrayView<const Point<dim, Number>>(
-                           all_points_ref.data() + ptr_q, n_q_points));
-
-            buffer.reinit(n_dofs_per_cell);
-            diagonal_local.reinit(n_dofs_per_cell);
-            all_value_m.resize(n_q_points * N);
-            all_value_p.resize(n_q_points * N);
-
-            for (unsigned int i = 0; i < n_dofs_per_cell; ++i)
-              {
-                for (unsigned int j = 0; j < n_dofs_per_cell; ++j)
-                  buffer[j] = static_cast<Number>(i == j);
-
-                local_evaluate(buffer, ptr_q, 1, all_value_m.data());
-
-                buffer.reinit(n_dofs_per_cell);
-                local_integrate(
-                  buffer, ptr_q, 1, all_value_m.data(), all_value_p.data());
-
-                diagonal_local[i] = buffer[i];
-              }
-
-            const auto local_dofs = this->get_dof_indices(cell);
-            constraints.distribute_local_to_global(diagonal_local,
-                                                   local_dofs,
-                                                   diagonal);
-
-            ptr_q += n_q_points;
-          }
-
-  diagonal.compress(VectorOperation::add);
-}
-
-template <int dim, typename Number>
-void
-CouplingOperatorBase<dim, Number>::add_sparsity_pattern_entries(
-  SparsityPatternBase &dsp) const
-{
-  const auto constraints = &constraints_extended;
-
-  for (unsigned int i = 0; i < dof_indices.size(); i += n_dofs_per_cell)
-    {
-      std::vector<types::global_dof_index> a(dof_indices.begin() + i,
-                                             dof_indices.begin() + i +
-                                               n_dofs_per_cell);
-      std::vector<types::global_dof_index> b(dof_indices_ghost.begin() + i,
-                                             dof_indices_ghost.begin() + i +
-                                               n_dofs_per_cell);
-
-      constraints->add_entries_local_to_global(a, b, dsp);
-      constraints->add_entries_local_to_global(b, a, dsp);
-    }
-}
-
-template <int dim, typename Number>
-void
-CouplingOperatorBase<dim, Number>::add_system_matrix_entries(
-  TrilinosWrappers::SparseMatrix &system_matrix) const
-{
-  const auto constraints = &constraints_extended;
-
-  std::vector<Number> all_value_m(all_normals.size() * n_dofs_per_cell * N);
-  std::vector<Number> all_value_p(all_normals.size() * n_dofs_per_cell * N);
-
-  unsigned int ptr_q = 0;
-
-  Vector<Number> buffer;
-
-  for (const auto &cell : dof_handler.active_cell_iterators())
-    if (cell->is_locally_owned())
-      for (const auto &face : cell->face_iterators())
-        if ((face->boundary_id() == bid_0) || (face->boundary_id() == bid_1))
-          {
-            const unsigned int n_q_points =
-              mortar_manager_q->get_n_points(get_rad(cell, face));
-
-            local_reinit(cell,
-                         ArrayView<const Point<dim, Number>>(
-                           all_points_ref.data() + ptr_q, n_q_points));
-
-            buffer.reinit(n_dofs_per_cell);
-
-            for (unsigned int i = 0; i < n_dofs_per_cell; ++i)
-              {
-                for (unsigned int j = 0; j < n_dofs_per_cell; ++j)
-                  buffer[j] = static_cast<Number>(i == j);
-
-                local_evaluate(buffer,
-                               ptr_q,
-                               n_dofs_per_cell,
-                               all_value_m.data() +
-                                 (ptr_q * n_dofs_per_cell + i) * N);
-              }
-
-            ptr_q += n_q_points;
-          }
-
-  const unsigned n_q_points =
-    Utilities::pow(quadrature.get_tensor_basis()[0].size(), dim - 1);
-
-  partitioner_cell.template export_to_ghosted_array<Number, 0>(
-    ArrayView<const Number>(reinterpret_cast<Number *>(all_value_m.data()),
-                            all_value_m.size() * n_components),
-    ArrayView<Number>(reinterpret_cast<Number *>(all_value_p.data()),
-                      all_value_p.size() * n_components),
-    n_dofs_per_cell * n_q_points * N * n_components);
-
-
-  ptr_q                 = 0;
-  unsigned int ptr_dofs = 0;
-
-  for (const auto &cell : dof_handler.active_cell_iterators())
-    if (cell->is_locally_owned())
-      for (const auto &face : cell->face_iterators())
-        if ((face->boundary_id() == bid_0) || (face->boundary_id() == bid_1))
-          {
-            const unsigned int n_sub_cells =
-              mortar_manager_cell->get_n_points(get_rad(cell, face));
-
-            for (unsigned int sc = 0; sc < n_sub_cells; ++sc)
-              {
-                const unsigned int n_q_points =
-                  mortar_manager_q->get_n_points(get_rad(cell, face)) /
-                  n_sub_cells;
-
-                local_reinit(cell,
-                             ArrayView<const Point<dim, Number>>(
-                               all_points_ref.data() + ptr_q, n_q_points));
-
-                for (unsigned int bb = 0; bb < 2; ++bb)
-                  {
-                    FullMatrix<Number> cell_matrix(n_dofs_per_cell,
-                                                   n_dofs_per_cell);
-
-                    for (unsigned int i = 0; i < n_dofs_per_cell; ++i)
-                      {
-                        buffer.reinit(n_dofs_per_cell);
-                        if (bb == 0)
-                          local_integrate(buffer,
-                                          ptr_q,
-                                          n_dofs_per_cell,
-                                          all_value_m.data() +
-                                            (ptr_q * n_dofs_per_cell + i) * N,
-                                          nullptr);
-                        else
-                          local_integrate(buffer,
-                                          ptr_q,
-                                          n_dofs_per_cell,
-                                          nullptr,
-                                          all_value_p.data() +
-                                            (ptr_q * n_dofs_per_cell + i) * N);
-
-                        for (unsigned int j = 0; j < n_dofs_per_cell; ++j)
-                          cell_matrix[j][i] = buffer[j];
-                      }
-
-
-                    std::vector<types::global_dof_index> a(
-                      dof_indices.begin() + ptr_dofs,
-                      dof_indices.begin() + ptr_dofs + n_dofs_per_cell);
-
-                    if (bb == 0)
-                      {
-                        constraints->distribute_local_to_global(cell_matrix,
-                                                                a,
-                                                                system_matrix);
-                      }
-                    else
-                      {
-                        std::vector<types::global_dof_index> b(
-                          dof_indices_ghost.begin() + ptr_dofs,
-                          dof_indices_ghost.begin() + ptr_dofs +
-                            n_dofs_per_cell);
-
-                        constraints->distribute_local_to_global(cell_matrix,
-                                                                a,
-                                                                b,
-                                                                system_matrix);
-                      }
-                  }
-
-                ptr_dofs += n_dofs_per_cell;
-
-                ptr_q += n_q_points;
-              }
-          }
-
-  AssertDimension(ptr_q, all_normals.size());
-  AssertDimension(ptr_dofs, dof_indices.size());
-}
 
 #endif
